Mp3 (Implementing singleton method on mp3, dependency injection in MusicPlaylistServiceImpl

+++ src/MusicPlayer/services/MusicPlaylistServicesImpl.java
@@ -1,23 +1,38 @@
 package MusicPlayer.services;
 
+import MusicPlayer.entity.Mp3;
 import MusicPlayer.entity.Music;
+import MusicPlayer.exception.Mp3Exception;
 import MusicPlayer.repository.PlaylistDB;
 
 public class MusicPlaylistServicesImpl  implements MusicPlaylistServices{
 
-    private PlaylistDB playlistDB;
+    private final PlaylistDB playlistDB;
+    private final Mp3 mp3;
 
-    public MusicPlaylistServicesImpl() {
-        playlistDB = PlaylistDB.getInstance();
+    public MusicPlaylistServicesImpl(PlaylistDB playlistDB) {
+        this.playlistDB = playlistDB;
+        mp3 = new Mp3();
+    }
+
+    public PlaylistDB getPlaylistDB() {
+        return playlistDB;
     }
 
     @Override
-    public void addToPlaylist(Music songTitle) {
-            playlistDB.download(songTitle);
+    public void addToPlaylist(Music songTitle) throws Mp3Exception {
+        if(mp3.isOn()) {
+            throw new Mp3Exception("Mp3 is off");
+        }
+        playlistDB.download(songTitle);
+
     }
 
     @Override
-    public void delete(Music songTitle) {
+    public void delete(String songTitle) throws Mp3Exception {
+        if(mp3.isOn()) {
+            throw new Mp3Exception("Mp3 is off");
+        }
         playlistDB.deleteMusic(songTitle);
     }
 
@@ -27,7 +42,7 @@ public class MusicPlaylistServicesImpl  implements MusicPlaylistServices{
     }
 
     @Override
-    public Music search(Music songTitle) {
+    public Music search(String songTitle) throws Mp3Exception {
         return playlistDB.searchFor(songTitle);
     }
